clean: extract twemproxy resource from backend
