+++ pkg/generators/backend/listener_deployment.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/3scale/saas-operator/pkg/resource_builders/marin3r"
 	"github.com/3scale/saas-operator/pkg/resource_builders/pod"
+	"github.com/3scale/saas-operator/pkg/resource_builders/twemproxy"
 	"github.com/3scale/saas-operator/pkg/util"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -68,7 +69,7 @@ func (gen *ListenerGenerator) deployment() func() *appsv1.Deployment {
 		}
 
 		if gen.TwemproxySpec != nil {
-			dep = addTwemproxySidecar(*dep, gen.TwemproxySpec)
+			dep = twemproxy.AddTwemproxySidecar(*dep, gen.TwemproxySpec)
 		}
 
 		return dep
